+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0
 
-- Initial experimental release.
+- Initial experimental release, based on stylelint-checkstyle-formatter